feate : 대시보드에 최근 일주일 감정 스코어 그래프 출력

+++ domain/src/main/java/com/songjem/domain/model/DashBoardEmotionItem.kt
@@ -2,6 +2,6 @@ package com.songjem.domain.model
 
 data class DashBoardEmotionItem(
     val targetDate: String,
-    val emotionScore: Int,
-    val neutralScore: Int
+    val emotionScore: Float,
+    val neutralScore: Float
 )
